Testing non commercial backends accepts cancellation. (#690)
* Testing non commercial backends accepts cancellation.

* Passing linter

* Marking test_cancel of ibmqjobs as slow

* Factoring out job tests-specific methods.

* Adding changelog modification

* Increasing timeout

* Distinguishing backend in case the tests are passed with IBM-Q credentials
