+++ requirements.txt
@@ -1,4 +1,4 @@
-IBMQuantumExperience>=1.9.8
+IBMQuantumExperience>=2.0.1
 matplotlib>=2.1
 networkx>=2.0
 numpy>=1.13
